test: create new abstract robot for dialogs
